Refactor x402 MCP wrapper to x402 tool, update README and test script for improved usage instructions and error handling
