+++ src/tools/x402-mcp/README.md
@@ -1,6 +1,6 @@
-# x402 MCP Wrapper
+# x402 Tool
 
-A wrapper around the x402 payment protocol MCP server that enables making paid API requests to x402-enabled services.
+A LangChain tool that enables making paid API requests to x402-enabled services using the x402 payment protocol.
 
 ## Features
 
@@ -8,10 +8,12 @@ A wrapper around the x402 payment protocol MCP server that enables making paid A
 - ✅ **Multiple HTTP methods**: Supports GET, POST, PUT, DELETE, PATCH
 - ✅ **Automatic payment handling**: Automatically handles x402 payment protocol
 - ✅ **Flexible configuration**: Only requires a private key
+- ✅ **Direct integration**: Simple LangChain tool (no MCP server layer)
 
 ## Prerequisites
 
 1. **Install dependencies**:
+
    ```bash
    yarn add viem x402-axios
    ```
@@ -25,6 +27,7 @@ A wrapper around the x402 payment protocol MCP server that enables making paid A
 ### Option 1: Use the test script
 
 1. **Set up a test x402 server** (if you don't have one):
+
    ```bash
    # Clone the x402 repo
    git clone https://github.com/coinbase/x402.git
@@ -35,10 +38,15 @@ A wrapper around the x402 payment protocol MCP server that enables making paid A
    ```
 
 2. **Run the test**:
+
    ```bash
+   # Uses PRIVATE_KEY from character's .env file (blockchainConfig.PRIVATE_KEY)
+   yarn test-x402 <character_name>
+
+   # Or set it as an environment variable
    export PRIVATE_KEY=0xYourPrivateKey
    export TEST_URL=http://localhost:4021/weather  # Optional, defaults to localhost:4021/weather
-   yarn test-x402
+   yarn test-x402 <character_name>
    ```
 
 ### Option 2: Use in your agent
@@ -49,8 +57,8 @@ Add to `src/agent.ts`:
 import { createX402Tools } from './tools/x402-mcp/index.js';
 
 // In createAgentRunnerOptions:
-const x402Tools = config.X402_PRIVATE_KEY
-  ? await createX402Tools({ privateKey: config.X402_PRIVATE_KEY })
+const x402Tools = config.blockchainConfig?.PRIVATE_KEY
+  ? await createX402Tools({ privateKey: config.blockchainConfig.PRIVATE_KEY })
   : [];
 
 const tools = [
@@ -59,7 +67,7 @@ const tools = [
 ].filter(tool => tool !== undefined);
 ```
 
-Then add `X402_PRIVATE_KEY` to your character's config.
+The tool will use the `PRIVATE_KEY` from your character's `.env` file (from `blockchainConfig.PRIVATE_KEY`).
 
 ## Usage
 
@@ -98,12 +106,12 @@ await tool.invoke({
 
 ## How It Works
 
-1. The wrapper creates an MCP server that uses `x402-axios` to intercept HTTP requests
+1. The tool uses `x402-axios` to create an axios client with a payment interceptor
 2. When a request is made to an x402-enabled endpoint:
    - If the endpoint requires payment (HTTP 402), the interceptor handles the payment automatically
    - Payment is made using the provided private key
    - The request is retried after payment
-3. The response is returned to the caller
+3. The response is returned to the caller as a JSON string
 
 ## Security Notes
 
@@ -111,4 +119,3 @@ await tool.invoke({
 - ⚠️ Use environment variables or secure config management
 - ⚠️ For production, use a dedicated wallet with limited funds
 - ⚠️ Test on Base Sepolia testnet first
-
