+++ src/tools/x402-mcp/manifest.json
@@ -1,16 +1,15 @@
 {
-  "name": "x402-mcp",
+  "name": "x402-tool",
   "version": "1.0.0",
   "description": "Tools for accessing paid APIs via x402 payment protocol",
   "author": "Jeremy Frank",
   "main": "index.ts",
   "dependencies": {
     "@langchain/core": "0.3.55",
-    "@autonomys/agent-core": "^0.3.1",
-    "@modelcontextprotocol/sdk": "^1.0.0",
     "axios": "^1.7.7",
     "viem": "^2.0.0",
-    "x402-axios": "^0.1.0"
+    "x402-axios": "^0.1.0",
+    "zod": "^3.23.8"
   },
-  "keywords": ["x402", "payment", "mcp", "api"]
+  "keywords": ["x402", "payment", "api"]
 }
